move funcs calling to and parsing output from entrez to a separate script
